LKE-14101: Bundle as a ESM module
