--- index.ts
@@ -1,39 +1,0 @@
-/**
- * Copyright Linkurious SAS 2012 - 2019
- *
- * - Created on 2016-11-28.
- */
-
-export * from './src/index';
-export * from './src/api/AccessRight';
-export * from './src/api/Alert';
-export * from './src/api/Application';
-export * from './src/api/Auth';
-export * from './src/api/Config';
-export * from './src/api/CustomAction';
-export * from './src/api/DataSource';
-export * from './src/api/favorite';
-export * from './src/api/Search';
-export * from './src/api/GraphEdge';
-export * from './src/api/GraphNode';
-export * from './src/api/GraphQuery';
-export * from './src/api/GraphSchema';
-export * from './src/api/import';
-export * from './src/api/License';
-export * from './src/api/Linkurious';
-export * from './src/api/Plugin';
-export * from './src/api/User';
-export * from './src/api/Visualization';
-export * from './src/api/webhook';
-export * from './src/api/commonTypes';
-export * from './src/api/displayTypes';
-export * from './src/api/spaces';
-export * from './src/api/graphItemTypes';
-export * from './src/api/nodeGrouping';
-export * from './src/http/request';
-export * from './src/http/response';
-export * from './src/http/types';
-export * from './src/index';
-export * from './src/errorListener';
-export * from './src/api/Tag';
-export * from './src/api/entityResolution';
