+++ package.json
@@ -1,13 +1,14 @@
 {
   "name": "@linkurious/rest-client",
   "version": "4.3.5",
-  "description": "A library to interact with Linkurious APIs",
+  "description": "A library to interact with Linkurious Enterprise's API",
   "files": [
-    "dist"
+    "dist/index.mjs",
+    "dist/index.mjs.map",
+    "dist/src/**/*.d.ts"
   ],
-  "main": "./dist/index.js",
-  "module": "./dist/index.js",
-  "typings": "./dist/index.d.ts",
+  "main": "./dist/index.mjs",
+  "types": "./dist/src/index.d.ts",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Linkurious/linkurious-rest-client.git"
@@ -22,26 +23,28 @@
   },
   "homepage": "https://github.com/Linkurious/linkurious-rest-client#readme",
   "scripts": {
-    "build": "tsc",
+    "build": "tsc && esbuild --bundle --sourcemap --target=es2020 --format=esm --outfile=dist/index.mjs src/index.ts",
     "test": "mocha dist/test",
     "test:unit": "mocha --reporter mocha-multi-reporters --reporter-options configFile=.test-reporters.json dist/test",
     "coverage": "nyc --reporter=text-summary --reporter=cobertura --report-dir=reports  mocha --reporter mocha-multi-reporters --reporter-options configFile=.test-reporters.json dist/test",
-    "compile": "tsc -b",
+    "compile": "tsc",
     "tsc": "tsc",
     "lint": "eslint --fix --ext .ts .",
     "lint:ci": "eslint --cache --cache-strategy content -f checkstyle -o reports/checkstyle.xml --ext .ts .",
     "watch": "tsc -w | tee",
-    "clean": "rm -rf node_modules; rm -rf package-lock.json; rm -rf dist",
+    "clean": "rm -rf node_modules dist reports .eslintcache .nyc_output",
     "bump:patch": "bump2version patch && npm version --no-git-tag-version patch",
     "bump:minor": "bump2version minor && npm version --no-git-tag-version minor",
     "bump:major": "bump2version major && npm version --no-git-tag-version --no-commit-hooks --no-workspaces-update major"
   },
   "devDependencies": {
-    "@types/mocha": "10.0.6",
     "@types/express": "4.17.15",
+    "@types/mocha": "10.0.6",
+    "@types/node": "24.12.0",
     "@types/superagent": "4.1.20",
     "@typescript-eslint/eslint-plugin": "7.15.0",
     "@typescript-eslint/parser": "7.15.0",
+    "esbuild": "0.28.0",
     "eslint": "8.57.0",
     "eslint-config-prettier": "9.1.0",
     "eslint-plugin-es5": "1.5.0",
@@ -52,10 +55,8 @@
     "mocha-multi-reporters": "1.5.1",
     "nyc": "15.1.0",
     "prettier": "3.1.1",
+    "superagent": "8.1.2",
     "typescript": "5.5.4"
-  },
-  "dependencies": {
-    "superagent": "8.1.2"
   },
   "publishConfig": {
     "access": "public"
